+++ packages/contracts-ethers/npm/events/PluginRepo.ts
@@ -1,0 +1,20 @@
+export {
+  GrantedEvent,
+  RevokedEvent,
+  UpgradedEvent,
+  InitializedEvent,
+  AdminChangedEvent,
+  BeaconUpgradedEvent,
+  VersionCreatedEvent,
+  ReleaseMetadataUpdatedEvent,
+} from '../../types/framework/plugin/repo/PluginRepo';
+
+export type EventName =
+  | 'Granted'
+  | 'Revoked'
+  | 'Upgraded'
+  | 'Initialized'
+  | 'AdminChanged'
+  | 'BeaconUpgraded'
+  | 'VersionCreated'
+  | 'ReleaseMetadataUpdated';