+++ packages/contracts-ethers/npm/events/PluginRepoRegistry.ts
@@ -1,0 +1,14 @@
+export {
+  UpgradedEvent,
+  InitializedEvent,
+  AdminChangedEvent,
+  BeaconUpgradedEvent,
+  PluginRepoRegisteredEvent,
+} from '../../types/framework/plugin/repo/PluginRepoRegistry';
+
+export type EventName =
+  | 'Upgraded'
+  | 'Initialized'
+  | 'AdminChanged'
+  | 'BeaconUpgraded'
+  | 'PluginRepoRegistered';