+++ packages/contracts-ethers/npm/events/PluginSetupProcessor.ts
@@ -1,0 +1,16 @@
+export {
+  UpdateAppliedEvent,
+  UpdatePreparedEvent,
+  InstallationAppliedEvent,
+  InstallationPreparedEvent,
+  UninstallationAppliedEvent,
+  UninstallationPreparedEvent,
+} from '../../types/framework/plugin/setup/PluginSetupProcessor';
+
+export type EventName =
+  | 'UpdateApplied'
+  | 'UpdatePrepared'
+  | 'InstallationApplied'
+  | 'InstallationPrepared'
+  | 'UninstallationApplied'
+  | 'UninstallationPrepared';