+++ packages/contracts-ethers/npm/events/index.ts
@@ -1,0 +1,5 @@
+export * as DAOEvents from './DAO';
+export * as DAORegistryEvents from './DAORegistry';
+export * as PluginRepoEvents from './PluginRepo';
+export * as PluginRepoRegistryEvents from './PluginRepoRegistry';
+export * as PluginSetupProcessorEvents from './PluginSetupProcessor';