+++ packages/contracts-ethers/npm/structs/PluginSetupProcessor.ts
@@ -1,0 +1,26 @@
+import {
+  PluginSetupProcessor,
+  PluginSetupRefStruct as PluginSetupRefStruct_,
+  PluginSetupRefStructOutput as PluginSetupRefStructOutput_,
+} from '../../types/framework/plugin/setup/PluginSetupProcessor';
+
+export type ApplyInstallationParamsStruct =
+  PluginSetupProcessor.ApplyInstallationParamsStruct;
+
+export type ApplyUninstallationParamsStruct =
+  PluginSetupProcessor.ApplyUninstallationParamsStruct;
+
+export type ApplyUpdateParamsStruct =
+  PluginSetupProcessor.ApplyUpdateParamsStruct;
+
+export type PrepareInstallationParamsStruct =
+  PluginSetupProcessor.PrepareInstallationParamsStruct;
+export type PrepareUninstallationParamsStruct =
+  PluginSetupProcessor.PrepareUninstallationParamsStruct;
+
+export type PrepareUpdateParamsStructOutput =
+  PluginSetupProcessor.PrepareUpdateParamsStructOutput;
+
+export type PluginSetupRefStruct = PluginSetupRefStruct_;
+
+export type PluginSetupRefStructOutput = PluginSetupRefStructOutput_;