+++ packages/contracts-ethers/npm/structs/index.ts
@@ -1,0 +1,4 @@
+export * as DAOStructs from './DAO';
+export * as DAOFactoryStructs from './DAOFactory';
+export * as PluginRepoStructs from './PluginRepo';
+export * as PluginSetupProcessorStructs from './PluginSetupProcessor';